% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{create_study}
\alias{create_study}
\title{Create a new study builder}
\usage{
create_study(study_id = "STUDY001")
}
\arguments{
\item{study_id}{Study identifier}
}
\description{
Create a new study builder
}
