% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{generate_complete_longitudinal_study}
\alias{generate_complete_longitudinal_study}
\title{Generate Complete Longitudinal Study with Full Pipeline Processing}
\usage{
generate_complete_longitudinal_study(
  participant_count = 1000,
  site_count = 150,
  study_id = "AA-AA-000-0000",
  snapshot_count = 3,
  snapshot_width = "months",
  start_date = as.Date("2025-02-01"),
  core_mappings = c("AE", "COUNTRY", "DATACHG", "DATAENT", "ENROLL", "LB", "VISIT", "PD",
    "PK", "QUERY", "STUDY", "STUDCOMP", "SDRGCOMP", "SITE", "SUBJ", "IE", "EXCLUSION")
)
}
\arguments{
\item{participant_count}{Number of study participants (default: 1000)}

\item{site_count}{Number of study sites (default: 150)}

\item{study_id}{Study identifier (default: "AA-AA-000-0000")}

\item{snapshot_count}{Number of time points (default: 3)}

\item{snapshot_width}{Time between snapshots (default: "months")}

\item{start_date}{Study start date}

\item{core_mappings}{Vector of core mapping types to include}
}
\value{
List containing all pipeline results equivalent to manual workflow
}
\description{
This function directly addresses the workflow shown in issue #95 by providing
a single function call to replace the complex manual orchestration of multiple
packages and workflow steps.
}
\examples{
\dontrun{
# Direct replacement for the manual workflow in issue #95
results <- generate_complete_longitudinal_study(
  participant_count = 1000,
  site_count = 150,
  study_id = "AA-AA-000-0000",
  snapshot_count = 3,
  snapshot_width = "months"
)

# Access results similar to manual workflow
site_results <- results$reporting_site
country_results <- results$reporting_country
study_results <- results$reporting_study
}
}
