% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{generate_study_data}
\alias{generate_study_data}
\title{Generate study data with configuration}
\usage{
generate_study_data(
  config,
  workflow_path = "workflow/1_mappings",
  mappings = NULL,
  package = "gsm.mapping",
  verbose = FALSE
)
}
\arguments{
\item{config}{Study configuration object}

\item{workflow_path}{Path to workflow mappings (not used in new approach)}

\item{mappings}{Optional mappings list (not used in new approach)}

\item{package}{Package containing workflows (not used in new approach)}

\item{verbose}{Whether to print progress/output messages}
}
\value{
List of generated study data
}
\description{
Generate study data with configuration
}
