% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{generate_study_data_with_config}
\alias{generate_study_data_with_config}
\title{Generate study data using configuration object}
\usage{
generate_study_data_with_config(
  config,
  workflow_path = "workflow/1_mappings",
  mappings = NULL,
  package = "gsm.mapping"
)
}
\arguments{
\item{config}{StudyConfig object containing all configuration}

\item{workflow_path}{Path to workflow mappings}

\item{mappings}{Mapping specifications to use}

\item{package}{Package containing the workflows}
}
\value{
List of generated study data
}
\description{
Internal function to generate study data based on StudyConfig settings
}
\keyword{internal}
