% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.datasim-package.R
\docType{package}
\name{gsm.datasim-package}
\alias{gsm.datasim}
\alias{gsm.datasim-package}
\title{gsm.datasim: Synthetic Test Data Generator}
\description{
Generates synthetic clinical trial data for testing and development of clinical monitoring applications. Produces configurable multi-snapshot longitudinal datasets across SDTM-style and custom domains, with optional analytics and reporting pipelines powered by the 'workr' workflow engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-Public/gsm.datasim}
  \item \url{https://gilead-public.github.io/gsm.datasim}
  \item Report bugs at \url{https://github.com/Gilead-Public/gsm.datasim/issues}
}

}
\author{
\strong{Maintainer}: Laura Maxwell \email{laura.maxwell@atorusresearch.com}

Authors:
\itemize{
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
  \item Zelos Zhu \email{zelos.zhu@atorusresearch.com}
  \item Roman Rogoza \email{roman.rogoza@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
