% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{organize_analytics_results}
\alias{organize_analytics_results}
\title{Organize analytics results into intuitive structure}
\usage{
organize_analytics_results(pipeline_results)
}
\arguments{
\item{pipeline_results}{Results from pipeline execution}
}
\value{
Organized analytics results
}
\description{
Organize analytics results into intuitive structure
}
