% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_longitudinal_study.R
\name{print.multiple_longitudinal_studies}
\alias{print.multiple_longitudinal_studies}
\title{Print method for multiple longitudinal studies}
\usage{
\method{print}{multiple_longitudinal_studies}(x, ...)
}
\arguments{
\item{x}{A multiple_longitudinal_studies object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for multiple longitudinal studies
}
\examples{
\dontrun{
studies <- create_multiple_longitudinal_studies(
  study_names = c("TRIAL-001", "TRIAL-002"),
  participants = 50, sites = 5, snapshots = 2
)
print(studies)
}
}
