% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_ENROLL.R
\name{Raw_ENROLL}
\alias{Raw_ENROLL}
\title{Generate Raw ENROLL Data}
\usage{
Raw_ENROLL(data, previous_data, spec, ...)
}
\arguments{
\item{data}{List of data frames pertaining to a simulated study which is passed through in a list of snapshots}

\item{previous_data}{List of data frames pertaining to prior snpashot with relation to \code{data}}

\item{spec}{A list representing the combined specifications from \code{CombineSpecs} across the desired domains for a simulation.}

\item{...}{Additional arguments to be passed.}
}
\value{
a data.frame pertaining to the raw dataset plugged into \code{ENROLL.yaml}
}
\description{
Generate Raw ENROLL based on \code{ENROLL.yaml} from \code{gsm.mapping}.
}
\seealso{
Other internal: 
\code{\link{Raw_AE}()},
\code{\link{Raw_DATACHG}()},
\code{\link{Raw_DATAENT}()},
\code{\link{Raw_LB}()},
\code{\link{Raw_PD}()},
\code{\link{Raw_QUERY}()},
\code{\link{Raw_SDRGCOMP}()},
\code{\link{Raw_SITE}()},
\code{\link{Raw_STUDCOMP}()},
\code{\link{Raw_STUDY}()},
\code{\link{Raw_SUBJ}()},
\code{\link{Raw_SV}()},
\code{\link{Raw_VISIT}()}
}
\concept{internal}
\keyword{internal}
