% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_SITE.R
\name{Raw_SITE}
\alias{Raw_SITE}
\title{Generate Raw SITE Data}
\usage{
Raw_SITE(data, previous_data, spec, ...)
}
\arguments{
\item{data}{List of data frames pertaining to a simulated study which is passed through in a list of snapshots}

\item{previous_data}{List of data frames pertaining to prior snpashot with relation to \code{data}}

\item{spec}{A list representing the combined specifications from \code{CombineSpecs} across the desired domains for a simulation.}

\item{...}{Additional arguments to be passed.}
}
\value{
a data.frame pertaining to the raw dataset plugged into \code{SITE.yaml}
}
\description{
Generate Raw SITE based on \code{SITE.yaml} from \code{gsm.mapping}.
}
\seealso{
Other internal: 
\code{\link{Raw_AE}()},
\code{\link{Raw_DATACHG}()},
\code{\link{Raw_DATAENT}()},
\code{\link{Raw_ENROLL}()},
\code{\link{Raw_LB}()},
\code{\link{Raw_PD}()},
\code{\link{Raw_QUERY}()},
\code{\link{Raw_SDRGCOMP}()},
\code{\link{Raw_STUDCOMP}()},
\code{\link{Raw_STUDY}()},
\code{\link{Raw_SUBJ}()},
\code{\link{Raw_SV}()}
}
\concept{internal}
\keyword{internal}
