% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_SUBJ.R
\name{Raw_SUBJ}
\alias{Raw_SUBJ}
\title{Generate Raw SUBJ Data}
\usage{
Raw_SUBJ(data, previous_data, spec, startDate, endDate, ...)
}
\arguments{
\item{data}{List of data frames pertaining to a simulated study which is passed through in a list of snapshots}

\item{previous_data}{List of data frames pertaining to prior snpashot with relation to \code{data}}

\item{spec}{A list representing the combined specifications from \code{CombineSpecs} across the desired domains for a simulation.}

\item{startDate}{The beginning of dates to which the subjects can be enrolled/started}

\item{endDate}{The end of dates to which the subjects completed/leave study.}

\item{...}{Additional arguments to be passed.}
}
\value{
a data.frame pertaining to the raw dataset plugged into \code{SUBJ.yaml}
}
\description{
Generate Raw SUBJ Data based on \code{SUBJ.yaml} from \code{gsm.mapping}.
}
\seealso{
Other internal: 
\code{\link{Raw_AE}()},
\code{\link{Raw_DATACHG}()},
\code{\link{Raw_DATAENT}()},
\code{\link{Raw_ENROLL}()},
\code{\link{Raw_LB}()},
\code{\link{Raw_PD}()},
\code{\link{Raw_QUERY}()},
\code{\link{Raw_SDRGCOMP}()},
\code{\link{Raw_SITE}()},
\code{\link{Raw_STUDCOMP}()},
\code{\link{Raw_STUDY}()},
\code{\link{Raw_SV}()},
\code{\link{Raw_VISIT}()}
}
\concept{internal}
\keyword{internal}
