% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_VISIT.R
\name{Raw_VISIT}
\alias{Raw_VISIT}
\title{Generate Raw VISIT Data}
\usage{
Raw_VISIT(data, previous_data, spec, startDate, ...)
}
\arguments{
\item{startDate}{The beginning of dates to which the subjects' visits are}
}
\value{
a data.frame pertaining to the raw dataset plugged into \code{Visit.yaml}
}
\description{
Generate Raw VISIT Data based on \code{Visit.yaml} from \code{gsm.mapping}.
}
\concept{internal}
\keyword{internal}
