% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_testdata_bundle.R
\name{read_testdata_bundle}
\alias{read_testdata_bundle}
\title{Read a test data bundle from disk}
\usage{
read_testdata_bundle(path)
}
\arguments{
\item{path}{The bundle folder containing \code{manifest.json}.}
}
\value{
An object of class \code{testdata_bundle} with \code{path} and \code{manifest}.
}
\description{
Rehydrates a bundle written by \code{\link[=build_testdata_bundle]{build_testdata_bundle()}} (or downloaded from
a release) so it can be summarised or explored. The data itself is read
lazily by the consuming functions.
}
\examples{
\dontrun{
bundle <- read_testdata_bundle("~/gsm-testdata/AA-AA-000-0000")
summarize_testdata_bundle(bundle)
}
}
