% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{set_temporal_config}
\alias{set_temporal_config}
\title{Set Temporal Configuration}
\usage{
set_temporal_config(
  config,
  start_date = NULL,
  snapshot_count = NULL,
  snapshot_width = NULL,
  end_date = NULL
)
}
\arguments{
\item{config}{Study configuration list}

\item{start_date}{Study start date}

\item{snapshot_count}{Number of snapshots}

\item{snapshot_width}{Time between snapshots}

\item{end_date}{Study end date}
}
\value{
Updated study configuration
}
\description{
Updates temporal configuration settings in a study config.
}
\examples{
config <- create_study_config("TRIAL001")
config <- set_temporal_config(config, start_date = "2023-06-01", snapshot_count = 6)
config$temporal_config$snapshot_count
}
