% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{study_builder_examples}
\alias{study_builder_examples}
\title{Example of fluent interface usage}
\description{
Example of fluent interface usage
}
\examples{
\dontrun{
# Simple study with standard datasets
study_data <- create_study("ONCOLOGY001") \%>\%
  with_study_design(participants = 200, sites = 15) \%>\%
  over_time(start_date = "2023-01-01", snapshots = 12, frequency = "months") \%>\%
  with_standard_datasets(adverse_events = TRUE, subject_visits = TRUE, visit_schedule = TRUE) \%>\%
  with_adverse_events(rate_per_patient = 2.5, temporal_pattern = "increasing") \%>\%
  generate()

# Preview configuration before generation
create_study("TRIAL002") \%>\%
  with_study_design(participants = 100, sites = 10) \%>\%
  with_standard_datasets() \%>\%
  preview()

# Custom dataset configuration
study_data <- create_study("CUSTOM001") \%>\%
  with_study_design(participants = 300, sites = 20) \%>\%
  add_custom_dataset("Raw_Biomarker",
                    count = function(config) config$study_params$participant_count * 5,
                    depends_on = "Raw_SUBJ") \%>\%
  generate()
}
}
