% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{StudyConfig}
\alias{StudyConfig}
\title{Study Configuration Class}
\description{
Study Configuration Class

Study Configuration Class
}
\details{
Manages study parameters, temporal configuration, and dataset specifications
for clinical trial data generation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{study_params}}{List of basic study parameters}

\item{\code{temporal_config}}{List of temporal configuration settings}

\item{\code{dataset_configs}}{List of dataset configurations
Initialize a new StudyConfig}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyConfig-new}{\code{StudyConfig$new()}}
\item \href{#method-StudyConfig-set_temporal}{\code{StudyConfig$set_temporal()}}
\item \href{#method-StudyConfig-add_dataset}{\code{StudyConfig$add_dataset()}}
\item \href{#method-StudyConfig-remove_dataset}{\code{StudyConfig$remove_dataset()}}
\item \href{#method-StudyConfig-validate}{\code{StudyConfig$validate()}}
\item \href{#method-StudyConfig-clone}{\code{StudyConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$new(
  study_id = "STUDY001",
  participant_count = 100,
  site_count = 10,
  analytics_package = NULL,
  analytics_workflows = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_id}}{Study identifier}

\item{\code{participant_count}}{Number of participants}

\item{\code{site_count}}{Number of sites}

\item{\code{analytics_package}}{Analytics package to use}

\item{\code{analytics_workflows}}{Specific workflows to run
Set temporal configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-set_temporal"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-set_temporal}{}}}
\subsection{Method \code{set_temporal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$set_temporal(
  start_date = NULL,
  snapshot_count = NULL,
  snapshot_width = NULL,
  end_date = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{Study start date}

\item{\code{snapshot_count}}{Number of snapshots}

\item{\code{snapshot_width}}{Time between snapshots}

\item{\code{end_date}}{Study end date
Add a dataset configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-add_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-add_dataset}{}}}
\subsection{Method \code{add_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$add_dataset(
  dataset_type,
  enabled = TRUE,
  count_formula = NULL,
  growth_pattern = "linear",
  dependencies = character(0),
  custom_args = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_type}}{Type of dataset (e.g., "Raw_AE")}

\item{\code{enabled}}{Whether the dataset should be generated}

\item{\code{count_formula}}{Formula for calculating record count}

\item{\code{growth_pattern}}{How the dataset grows over time}

\item{\code{dependencies}}{Dependencies on other datasets}

\item{\code{custom_args}}{Additional arguments for the dataset generator
Remove a dataset configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-remove_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-remove_dataset}{}}}
\subsection{Method \code{remove_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$remove_dataset(dataset_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_type}}{Type of dataset to remove
Validate the configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-validate"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-validate}{}}}
\subsection{Method \code{validate()}}{
Basic validation of study parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
