% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{create_standard_study_config}
\alias{create_standard_study_config}
\title{Create Study Configuration for Standard Datasets}
\usage{
create_standard_study_config(
  study_id = "STUDY001",
  participant_count = 100,
  site_count = 10,
  analytics_package = NULL,
  analytics_workflows = NULL,
  study = TRUE,
  subjects = TRUE,
  sites_data = TRUE,
  adverse_events = TRUE,
  protocol_deviations = TRUE,
  lab_data = TRUE,
  subject_visits = TRUE,
  visit_schedule = TRUE,
  enrollment = TRUE,
  data_changes = TRUE,
  data_entry = TRUE,
  queries = TRUE,
  pharmacokinetics = TRUE,
  study_drug_completion = TRUE,
  study_completion = TRUE,
  inclusion_exclusion = TRUE,
  exclusions = TRUE,
  country = TRUE,
  death = TRUE,
  randomization = TRUE,
  overall_response = TRUE,
  outlier_intensity = 1
)
}
\arguments{
\item{study_id}{Study identifier}

\item{participant_count}{Number of participants}

\item{site_count}{Number of sites}

\item{analytics_package}{Analytics package to use}

\item{analytics_workflows}{Specific workflows to run}

\item{study}{Include study metadata (Raw_STUDY)}

\item{subjects}{Include subject demographics (Raw_SUBJ)}

\item{sites_data}{Include site information (Raw_SITE)}

\item{adverse_events}{Include adverse event data}

\item{protocol_deviations}{Include protocol deviation data}

\item{lab_data}{Include laboratory data}

\item{subject_visits}{Include subject visit data (Raw_VISIT)}

\item{visit_schedule}{Include visit schedule data (Raw_VISIT)}

\item{enrollment}{Include enrollment data}

\item{data_changes}{Include data change tracking (Raw_DATACHG)}

\item{data_entry}{Include data entry tracking (Raw_DATAENT)}

\item{queries}{Include query data (Raw_QUERY)}

\item{pharmacokinetics}{Include pharmacokinetics data}

\item{study_drug_completion}{Include study drug completion (Raw_SDRGCOMP)}

\item{study_completion}{Include overall study completion (Raw_STUDCOMP)}

\item{inclusion_exclusion}{Include inclusion/exclusion criteria (Raw_IE)}

\item{country}{Include country mapping}

\item{death}{Include death events (Raw_Death)}

\item{randomization}{Include randomization data (Raw_Randomization)}

\item{overall_response}{Include overall response data (Raw_OverallResponse)}

\item{outlier_intensity}{Global multiplier for outlier-like values in domain generators.}
}
\value{
Study configuration with standard datasets
}
\description{
Convenience function to create a study configuration with standard clinical datasets.
}
\examples{
# All default datasets
config <- create_standard_study_config("TRIAL001", participant_count = 100, site_count = 10)
names(config$dataset_configs)

# Select a subset of domains
config <- create_standard_study_config(
  "TRIAL002",
  participant_count = 50,
  adverse_events = TRUE, lab_data = TRUE,
  pharmacokinetics = FALSE, overall_response = FALSE
)
}
