% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{ensure_core_mappings}
\alias{ensure_core_mappings}
\title{Ensure core mappings are included with domains}
\usage{
ensure_core_mappings(domains)
}
\arguments{
\item{domains}{Vector of domain mapping names}
}
\value{
Vector of mapping names including required core mappings
}
\description{
Ensure core mappings are included with domains
}
