% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{ensure_core_mappings}
\alias{ensure_core_mappings}
\title{Ensure core mappings are included}
\usage{
ensure_core_mappings(domains)
}
\arguments{
\item{domains}{Vector of domain names}
}
\value{
Vector with required core mappings added
}
\description{
Ensure core mappings are included
}
