% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{execute_analytics_pipeline}
\alias{execute_analytics_pipeline}
\title{Execute the analytics pipeline}
\usage{
execute_analytics_pipeline(raw_data, config)
}
\arguments{
\item{raw_data}{List of raw data snapshots}

\item{config}{Study configuration}
}
\value{
Pipeline results
}
\description{
Execute the analytics pipeline
}
