% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{execute_analytics_pipeline}
\alias{execute_analytics_pipeline}
\title{Execute analytics pipeline}
\usage{
execute_analytics_pipeline(raw_data, config)
}
\arguments{
\item{raw_data}{Raw study data}

\item{config}{Study configuration object}
}
\value{
Analytics pipeline results
}
\description{
Execute analytics pipeline
}
