% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{execute_analytics_pipeline}
\alias{execute_analytics_pipeline}
\title{Execute analytics pipeline}
\usage{
execute_analytics_pipeline(raw_data, config)
}
\arguments{
\item{raw_data}{Raw study data}

\item{config}{Study configuration object}
}
\value{
Analytics pipeline results
}
\description{
Execute analytics pipeline
}
\examples{
\dontrun{
config <- create_standard_study_config("STUDY001", participant_count = 50, site_count = 5)
config <- set_temporal_config(config, snapshot_count = 1)
raw_data <- generate_study_data(config)
results <- execute_analytics_pipeline(raw_data, config)
}
}
