% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{generate_raw_data_for_endpoints}
\alias{generate_raw_data_for_endpoints}
\title{Generate raw data for endpoints study (multi-package)}
\usage{
generate_raw_data_for_endpoints(config, domain_package_df)
}
\arguments{
\item{config}{Study configuration object with enabled datasets}

\item{domain_package_df}{Data frame mapping domains to packages}
}
\value{
List of raw data for enabled datasets
}
\description{
Generate raw data for endpoints study (multi-package)
}
