% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_utils.R
\name{generate_raw_data_from_config}
\alias{generate_raw_data_from_config}
\title{Generate raw data for study configuration}
\usage{
generate_raw_data_from_config(config, verbose = FALSE)
}
\arguments{
\item{config}{Study configuration object with enabled datasets}

\item{verbose}{Whether to print progress/output messages}
}
\value{
List of raw data for enabled datasets
}
\description{
Generate raw data for study configuration
}
