% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rawdata_for_single_study.R
\name{generate_rawdata_for_single_study}
\alias{generate_rawdata_for_single_study}
\title{Generate Snapshots for Single Study Data}
\usage{
generate_rawdata_for_single_study(
  SnapshotCount,
  SnapshotWidth,
  ParticipantCount,
  SiteCount,
  StudyID,
  workflow_path,
  mappings,
  package,
  desired_specs = NULL
)
}
\arguments{
\item{SnapshotCount}{An integer specifying the number of snapshots to generate.}

\item{SnapshotWidth}{A character specifying the frequency of snapshots, defaults to "months".
Accepts "days", "weeks", "months" and "years". User can also place a number and unit such as "3 months".}

\item{ParticipantCount}{An integer specifying the number of participants in the study.}

\item{SiteCount}{An integer specifying the number of sites for the study.}

\item{StudyID}{A string specifying the study identifier.}

\item{workflow_path}{A string specifying the path to the workflow mappings.}

\item{mappings}{A string specifying the names of the workflows to run.}

\item{package}{A string specifying the package in which the workflows used in \code{MakeWorkflowList()} are located.}

\item{desired_specs}{A list of specifications of the data types that should be included.}
}
\value{
A list of data snapshots, where each element contains simulated data for a particular snapshot
period (typically a month), with variables populated according to the provided specifications.
}
\description{
This function generates a list of study data snapshots based on the provided specifications,
including participant count, site count, study ID, and the number of snapshots to be generated.
The data is simulated over a series of months, with key variables such as subject enrollment,
adverse events, and time on study being populated according to the study's specifications.
}
\details{
The function generates snapshots over a sequence of months,starting from \code{"2012-01-01"}. For each snapshot:
\enumerate{
\item The number of adverse events (\code{ae_num}) is simulated.
\item The number of participants screened is determined.
\item Key variables (such as subject ID, enrollment date, time on study, etc.) are populated for each data type
specified in \code{combined_specs}.
}
}
\examples{
snapshots <- generate_rawdata_for_single_study(
  SnapshotCount = 3,
  SnapshotWidth = "months",
  ParticipantCount = 50,
  SiteCount = 5,
  StudyID = "ABC",
  workflow_path = "workflow/1_mappings",
  mappings = "AE",
  package = "gsm.mapping",
  desired_specs = NULL
)

}
