% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_utils.R
\name{generate_study_data_with_config}
\alias{generate_study_data_with_config}
\title{Generate study data with configuration}
\usage{
generate_study_data_with_config(
  config,
  workflow_path = "workflow/1_mappings",
  mappings = NULL,
  package = "gsm.mapping"
)
}
\arguments{
\item{config}{StudyConfig object}

\item{workflow_path}{Path to workflow mappings (not used in new approach)}

\item{mappings}{Optional mappings list (not used in new approach)}

\item{package}{Package containing workflows (not used in new approach)}
}
\value{
List of generated study data
}
\description{
Generate study data with configuration
}
