% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_utils.R
\name{generate_study_snapshots}
\alias{generate_study_snapshots}
\title{Generate study data across multiple snapshots}
\usage{
generate_study_snapshots(
  study_id,
  participants,
  sites,
  snapshots,
  interval,
  mappings,
  base_date = NULL,
  verbose = FALSE
)
}
\arguments{
\item{study_id}{Study identifier}

\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{snapshots}{Number of snapshots}

\item{interval}{Time interval between snapshots}

\item{mappings}{Vector of mapping names to use}

\item{base_date}{Base date for snapshot generation (defaults to "2012-01-31" if NULL)}

\item{verbose}{Whether to print progress/output messages}
}
\value{
List of raw data for each snapshot
}
\description{
Generate study data across multiple snapshots
}
