% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_utils.R
\name{generate_study_snapshots}
\alias{generate_study_snapshots}
\title{Generate study data across multiple timepoints}
\usage{
generate_study_snapshots(
  study_id,
  participants,
  sites,
  timepoints,
  interval,
  mappings
)
}
\arguments{
\item{study_id}{Study identifier}

\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{timepoints}{Number of timepoints}

\item{interval}{Time interval between snapshots}

\item{mappings}{Vector of mapping names to use}
}
\value{
List of raw data for each timepoint
}
\description{
Generate study data across multiple timepoints
}
