% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{generate_study_snapshots}
\alias{generate_study_snapshots}
\title{Generate study data snapshots}
\usage{
generate_study_snapshots(
  study_id,
  participants,
  sites,
  timepoints,
  interval,
  mappings
)
}
\arguments{
\item{study_id}{Study identifier}

\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{timepoints}{Number of timepoints}

\item{interval}{Time interval}

\item{mappings}{Technical mapping names}
}
\value{
List of raw data snapshots
}
\description{
Generate study data snapshots
}
\keyword{internal}
