% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{generate_study_snapshots}
\alias{generate_study_snapshots}
\title{Generate study data across multiple snapshots}
\usage{
generate_study_snapshots(
  study_id,
  participants,
  sites,
  snapshots,
  interval,
  mappings,
  base_date = NULL,
  outlier_intensity = 1,
  verbose = FALSE
)
}
\arguments{
\item{study_id}{Study identifier}

\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{snapshots}{Number of snapshots}

\item{interval}{Time interval between snapshots}

\item{mappings}{Vector of mapping names to use}

\item{base_date}{Base date for snapshot generation (defaults to "2012-01-31" if NULL)}

\item{outlier_intensity}{Global multiplier for outlier-like values in domain generators.}

\item{verbose}{Whether to print progress/output messages}
}
\value{
List of raw data for each snapshot
}
\description{
Generate study data across multiple snapshots
}
\examples{
\dontrun{
mappings <- ensure_core_mappings(c("AE", "LB"))
snapshots <- generate_study_snapshots(
  study_id = "STUDY-001",
  participants = 100, sites = 10, snapshots = 3,
  interval = "1 month", mappings = mappings
)
length(snapshots)
}
}
