% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.datasim-package.R
\docType{package}
\name{gsm.datasim-package}
\alias{gsm.datasim}
\alias{gsm.datasim-package}
\title{gsm.datasim: Synthetic Test Data Generator}
\description{
Generates synthetic test data for the RBQM of Clinical Trials based on several parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-biostats.github.io/gsm.datasim/}
}

}
\author{
\strong{Maintainer}: Roman Rogoza \email{roman.rogoza@atorusresearch.com}

Authors:
\itemize{
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
}

}
\keyword{internal}
