% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{LongitudinalStudy}
\alias{LongitudinalStudy}
\title{Longitudinal Study Data Container}
\description{
Longitudinal Study Data Container

Longitudinal Study Data Container
}
\details{
R6 class that encapsulates longitudinal study data and provides intuitive
access methods for different analysis perspectives.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{study_id}}{Study identifier}

\item{\code{raw_data}}{List of raw data snapshots}

\item{\code{config}}{Study configuration parameters}

\item{\code{analytics}}{Processed analytics results
Initialize new LongitudinalStudy}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LongitudinalStudy-new}{\code{LongitudinalStudy$new()}}
\item \href{#method-LongitudinalStudy-summary}{\code{LongitudinalStudy$summary()}}
\item \href{#method-LongitudinalStudy-run_analytics_pipeline}{\code{LongitudinalStudy$run_analytics_pipeline()}}
\item \href{#method-LongitudinalStudy-get_timepoint}{\code{LongitudinalStudy$get_timepoint()}}
\item \href{#method-LongitudinalStudy-get_domain_timeline}{\code{LongitudinalStudy$get_domain_timeline()}}
\item \href{#method-LongitudinalStudy-get_domain_names}{\code{LongitudinalStudy$get_domain_names()}}
\item \href{#method-LongitudinalStudy-clone}{\code{LongitudinalStudy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$new(study_id, raw_data, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_id}}{Study identifier}

\item{\code{raw_data}}{Raw study data snapshots}

\item{\code{config}}{Configuration parameters
Get summary of study structure}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-summary"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-summary}{}}}
\subsection{Method \code{summary()}}{
Display comprehensive study summary
Run the complete analytics pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-run_analytics_pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-run_analytics_pipeline}{}}}
\subsection{Method \code{run_analytics_pipeline()}}{
Execute full analytics pipeline on study data
Get data for specific timepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$run_analytics_pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-get_timepoint"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-get_timepoint}{}}}
\subsection{Method \code{get_timepoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$get_timepoint(timepoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timepoint}}{Timepoint number (1-based)
Get specific domain data across all timepoints}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-get_domain_timeline"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-get_domain_timeline}{}}}
\subsection{Method \code{get_domain_timeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$get_domain_timeline(domain_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain_name}}{Domain mapping name (e.g., "AE", "LB")
Get available domain names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-get_domain_names"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-get_domain_names}{}}}
\subsection{Method \code{get_domain_names()}}{
Return list of available domain names across all timepoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$get_domain_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LongitudinalStudy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LongitudinalStudy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LongitudinalStudy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
