% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_utils.R
\name{organize_analytics_results}
\alias{organize_analytics_results}
\title{Organize analytics pipeline results}
\usage{
organize_analytics_results(pipeline_results, verbose = FALSE)
}
\arguments{
\item{pipeline_results}{Raw pipeline results from gsm.core}

\item{verbose}{Whether to print progress/output messages}
}
\value{
Organized analytics results
}
\description{
Organize analytics pipeline results
}
