% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R, R/study_utils.R
\name{parse_interval_to_snapshot_width}
\alias{parse_interval_to_snapshot_width}
\title{Parse interval string to snapshot width}
\usage{
parse_interval_to_snapshot_width(interval)

parse_interval_to_snapshot_width(interval)
}
\arguments{
\item{interval}{Interval string (e.g., "1 month", "2 weeks")}
}
\value{
Snapshot width for temporal configuration

Snapshot width for temporal configuration
}
\description{
Parse interval string to snapshot width

Parse interval string to snapshot width
}
\examples{
parse_interval_to_snapshot_width("1 month")
parse_interval_to_snapshot_width("2 weeks")
parse_interval_to_snapshot_width("30 days")
}
