% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_generator.R
\name{raw_data_generator}
\alias{raw_data_generator}
\title{Generate Raw Data for Study Snapshots}
\usage{
raw_data_generator(
  ParticipantCount = NULL,
  SiteCount = NULL,
  StudyID = NULL,
  SnapshotCount = NULL,
  SnapshotWidth = NULL,
  template_path = "~/gsm.datasim/inst/template.csv",
  workflow_path = "workflow/1_mappings",
  generate_reports = FALSE,
  mappings = NULL,
  package = "gsm.mapping",
  save = FALSE
)
}
\arguments{
\item{ParticipantCount}{An integer specifying the number of participants for the study.
If \code{NULL}, the function will use values from the template file.}

\item{SiteCount}{An integer specifying the number of sites for the study. If \code{NULL}, the function
will use values from the template file.}

\item{StudyID}{A string specifying the study identifier. If \code{NULL}, the function will use values
from the template file.}

\item{SnapshotCount}{An integer specifying the number of snapshots for the study. If \code{NULL},
the function will use values from the template file.}

\item{SnapshotWidth}{A character specifying the frequency of snapshots, defaults to "months".
Accepts "days", "weeks", "months" and "years". User can also place a number and unit such as "3 months".}

\item{template_path}{A string specifying the path to the template CSV file. Default is
\code{"~/gsm.datasim/inst/template.csv"}.}

\item{workflow_path}{A string specifying the path to the workflow mappings. Default is
\code{"workflow/1_mappings"}.}

\item{generate_reports}{A boolean, specifying whether or not to produce reports upon execution. Default is FALSE.}

\item{mappings}{A string specifying the names of the workflows to run.}

\item{package}{A string specifying the package in which the workflows used in \code{MakeWorkflowList()} are located. Default is "gsm".}

\item{save}{A boolean, specifying whether or not this should be saved out as an RDS}
}
\value{
A list of raw data generated for each study snapshot, saved as an RDS file in \code{"data-raw/raw_data.RDS"}.
}
\description{
This function generates raw data for study snapshots based on provided participant count,
site count, study ID, and snapshot count. If any of these parameters are \code{NULL}, it loads
a template CSV file to iterate over multiple study configurations. The generated raw data
is saved as an RDS file.
}
\details{
The function performs the following steps:
\enumerate{
\item If \code{ParticipantCount}, \code{SiteCount}, \code{StudyID}, or \code{SnapshotCount} is \code{NULL}, the function reads
the \code{template.csv} file to get the necessary parameters for multiple studies.
\item It generates raw data for study snapshots based on either provided parameters or the template file.
\item The generated data is saved as an RDS file and returned as a list.
}
}
\examples{
\dontrun{
# Generate raw data using specified parameters
data <- raw_data_generator(ParticipantCount = 100, SiteCount = 10, StudyID = "Study01", SnapshotCount = 5, SnapshotWidth = "months")


# Generate raw data using a template file
data <- raw_data_generator()
}

}
