% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_longitudinal_study.R
\name{summary.multiple_longitudinal_studies}
\alias{summary.multiple_longitudinal_studies}
\title{Summary method for multiple longitudinal studies}
\usage{
\method{summary}{multiple_longitudinal_studies}(object, ...)
}
\arguments{
\item{object}{A multiple_longitudinal_studies object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for multiple longitudinal studies
}
