% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_builder.R
\name{validate_study_inputs}
\alias{validate_study_inputs}
\title{Validate study input parameters}
\usage{
validate_study_inputs(participants, sites, timepoints, domains)
}
\arguments{
\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{timepoints}{Number of timepoints}

\item{domains}{Domain specifications}
}
\description{
Validate study input parameters
}
