% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm_utils.R
\name{validate_study_inputs}
\alias{validate_study_inputs}
\title{Study Utility Functions}
\usage{
validate_study_inputs(participants, sites, snapshots, domains)
}
\arguments{
\item{participants}{Number of participants}

\item{sites}{Number of sites}

\item{snapshots}{Number of snapshots}

\item{domains}{Domains to include}
}
\description{
This file contains utility functions for study validation, data generation,
and analytics pipeline execution.
Validate study input parameters
}
